Antimony-Lead-Alloy-Market-Global-Outlook-and-Forecast-2024-2030.md
The global Antimony Lead Alloy market was valued at US$ million in 2023 and is projected to reach US$ million by 2030, at a CAGR of % during the forecast period. The influence of COVID-19 and the Russia-Ukraine War were considered while estimating market sizes. The U.S. Market is Estimated at $ Million in 2023, While China is Forecast to Reach $ Million. Antimony content